Introduce new and edit ifql connection forms
